requirements.txt
matplotlib
numpy
tensorflow
opencv-python
pandas
scikit-learn
plotly
nbformat
h5py
tqdm
opencv-python
sklearn-genetic-opt